requirements.txt
requests
pyspark
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
kafka-python-ng
lz4
rich
